4936 - Remove Country_summary materialized view
