--- src/server/repository/assessmentCycle/countrySummary/dropMaterializedView.ts
@@ -1,17 +1,0 @@
-import { Assessment } from 'meta/assessment/assessment'
-import { Cycle } from 'meta/assessment/cycle'
-
-import { BaseProtocol, DB, Schemas } from 'server/db'
-
-type Props = {
-  assessment: Assessment
-  cycle: Cycle
-}
-
-export const dropMaterializedView = async (props: Props, client: BaseProtocol = DB): Promise<void> => {
-  const { assessment, cycle } = props
-
-  const schemaCycle = Schemas.getNameCycle(assessment, cycle)
-
-  return client.query(`drop materialized view if exists ${schemaCycle}.country_summary`)
-}
