+++ src/server/repository/assessmentCycle/countrySummary/getCount.ts
@@ -4,7 +4,9 @@ import { Assessment } from 'meta/assessment/assessment'
 import { Cycle } from 'meta/assessment/cycle'
 import { TablePaginatedCount } from 'meta/tablePaginated'
 
-import { BaseProtocol, DB, Schemas } from 'server/db'
+import { BaseProtocol, DB } from 'server/db'
+
+import { getBaseQuery } from './_queries/getBaseQuery'
 
 type Props = {
   assessment: Assessment
@@ -14,12 +16,13 @@ type Props = {
 export const getCount = async (props: Props, client: BaseProtocol = DB): Promise<TablePaginatedCount> => {
   const { assessment, cycle } = props
 
-  const schemaCycle = Schemas.getNameCycle(assessment, cycle)
+  const baseQuery = getBaseQuery({ assessment, cycle })
 
   return client.one(
     `
-        select count(c.country_iso) as total
-        from ${schemaCycle}.country_summary c
+      ${baseQuery}
+      select count(cs.country_iso) as total
+      from country_summary cs
     `,
     [],
     (res) => Objects.camelize(res)
