+++ src/server/repository/assessmentCycle/countrySummary/getMany.ts
@@ -8,6 +8,7 @@ import { TablePaginatedOrderByDirection } from 'meta/tablePaginated'
 import { BaseProtocol, DB, Schemas } from 'server/db'
 import { SQLs } from 'server/db/SQLs'
 
+import { getBaseQuery } from './_queries/getBaseQuery'
 import { fields, fieldsFromCountryJoined } from './fields'
 
 type Props = {
@@ -23,15 +24,17 @@ export const getMany = async (props: Props, client: BaseProtocol = DB): Promise<
   const { assessment, cycle, limit, offset, orderBy, orderByDirection } = props
 
   const schemaCycle = Schemas.getNameCycle(assessment, cycle)
+  const baseQuery = getBaseQuery({ assessment, cycle })
 
   const query = `
-        select ${SQLs.fieldsJoined(fields, 'cs')}, ${fieldsFromCountryJoined('c')}, c.status
-        from ${schemaCycle}.country_summary cs
-            left join ${schemaCycle}.country c using (country_iso)
-        order by ${orderBy ?? 'country_iso'} ${orderByDirection ?? TablePaginatedOrderByDirection.asc} nulls last
-        ${limit ? 'limit $1' : ''} ${offset ? 'offset $2' : ''}
-        ;
-    `
+    ${baseQuery}
+    select ${SQLs.fieldsJoined(fields, 'cs')}, ${fieldsFromCountryJoined('c')}, c.status
+    from country_summary cs
+         left join ${schemaCycle}.country c using (country_iso)
+    order by ${orderBy ?? 'country_iso'} ${orderByDirection ?? TablePaginatedOrderByDirection.asc} nulls last
+    ${limit ? 'limit $1' : ''} ${offset ? 'offset $2' : ''}
+    ;
+  `
 
   const params = []
   if (limit) params.push(limit)
