+++ src/server/repository/assessmentCycle/countrySummary/getOneOrNone.ts
@@ -7,6 +7,7 @@ import { Cycle } from 'meta/assessment/cycle'
 import { BaseProtocol, DB, Schemas } from 'server/db'
 import { SQLs } from 'server/db/SQLs'
 
+import { getBaseQuery } from './_queries/getBaseQuery'
 import { fields } from './fields'
 
 type Props = {
@@ -19,13 +20,15 @@ export const getOneOrNone = async (props: Props, client: BaseProtocol = DB): Pro
   const { assessment, countryIso, cycle } = props
 
   const schemaCycle = Schemas.getNameCycle(assessment, cycle)
+  const baseQuery = getBaseQuery({ assessment, cycle })
 
   return client.oneOrNone<CountrySummary>(
     `
-        select ${SQLs.fieldsJoined(fields, 'cs')}, c.status
-        from ${schemaCycle}.country_summary cs
-                    left join ${schemaCycle}.country c using (country_iso)
-        where cs.country_iso = $1
+      ${baseQuery}
+      select ${SQLs.fieldsJoined(fields, 'cs')}, c.status
+      from country_summary cs
+           left join ${schemaCycle}.country c using (country_iso)
+      where cs.country_iso = $1
     `,
     [countryIso],
     Objects.camelize
